Move Header inside Home
